구조 변경: ContinueState → FlowState 이름 변경, 패키지 분리, OutputMessenger 통합
- ContinueState → FlowState 이름 변경 및 패키지 분리
  - 이유: FlowState가 더 많은 것을 담게 하려고 함, 아예 별도의 패키지로 분리

- OutputMessenger 통합
  - MenuGroupMessenger, MenuItemMessenger를 OutputMessenger로 통합
  - 이유: 서로 다른 Guide에서 사용하고 있긴 하지만, 기능이 많이 중복되어 코드가 중복됨
    - 인터페이스나 추상 클래스로 계층을 둘 수도 있겠으나, 아직은 그 필요성이 크지 않은 듯함
    - 당장의 관리 편의를 위해 통합, 따라서 필요한 경우 다시 분리하거나 계층을 나눌 수 있음
