--- week3/src/main/kotlin/kotlinassignment/week3/guide/ContinueState.kt
@@ -1,10 +1,0 @@
-package kotlinassignment.week3.guide
-
-import kotlinassignment.week3.menu.MenuGroup
-
-class ContinueState {
-
-    var nextGuide: Guide? = MenuGroupGuide()
-    var previousGuide: Guide = MenuGroupGuide()
-    lateinit var nextMenuGroup: MenuGroup
-}
