+++ week3/src/main/kotlin/kotlinassignment/week3/guide/Guide.kt
@@ -1,6 +1,8 @@
 package kotlinassignment.week3.guide
 
+import kotlinassignment.week3.flowState.FlowState
+
 interface Guide {
 
-    fun guide(continueState: ContinueState)
+    fun guide(flowState: FlowState)
 }
