+++ week3/src/main/kotlin/kotlinassignment/week3/guide/MenuGroupGuide.kt
@@ -1,30 +1,31 @@
 package kotlinassignment.week3.guide
 
+import kotlinassignment.week3.flowState.FlowState
 import kotlinassignment.week3.menu.MenuGroup
 import kotlinassignment.week3.messenger.InputMessenger
-import kotlinassignment.week3.messenger.MenuGroupMessenger
+import kotlinassignment.week3.messenger.OutputMessenger
 import kotlinassignment.week3.messenger.Message
 
 class MenuGroupGuide: Guide {
-    private val menuGroupMessenger = MenuGroupMessenger()
+    private val outputMessenger = OutputMessenger()
 
-    override fun guide(continueState: ContinueState) {
+    override fun guide(flowState: FlowState) {
         // 메뉴 출력에 대한 부분
         val menuGroupEntries = MenuGroup.entries
-        menuGroupMessenger.writeMenuGroup(menuGroupEntries)
+        outputMessenger.writeMenuGroup(menuGroupEntries)
         
         // 사용자의 입력 처리에 대한 부분
         val (inputStatus, selectedNumber) = InputMessenger().readInt()
         if (inputStatus == InputMessenger.InputStatus.ABNORMAL) return // 다시 menuGroupGuide의 guide를 호출해야하므로 nextGuide를 set하지 않음, 다음 명령어들을 실행하지 않기 위해 return
 
         if (selectedNumber == 0) {
-            menuGroupMessenger.write(Message.EXIT)
-            continueState.nextGuide = null
+            outputMessenger.write(Message.EXIT)
+            flowState.nextGuide = null
         } else if (selectedNumber <= menuGroupEntries.size) {
-            continueState.nextGuide = MenuItemGuide()
-            continueState.nextMenuGroup = menuGroupEntries[selectedNumber - 1]
+            flowState.nextGuide = MenuItemGuide()
+            flowState.nextMenuGroup = menuGroupEntries[selectedNumber - 1]
         } else {
-            menuGroupMessenger.write(Message.NO_CORRESPONDING_SERVICE_NUMBER)
+            outputMessenger.write(Message.NO_CORRESPONDING_SERVICE_NUMBER)
             // 다시 menuGroupGuide의 guide를 호출해야하므로 nextGuide를 set하지 않는다.
         }
     }
